Completed tasks for lesson2 task1 & task2

+++ src/lesson1/task1/Simple.kt
@@ -86,7 +86,7 @@ fun lengthInMeters(sagenes: Int, arshins: Int, vershoks: Int): Double =
  * Вывести значение того же угла в радианах (например, 0.63256).
  */
 fun angleInRadian(deg: Int, min: Int, sec: Int): Double =
-    (deg * 3600 + min * 60 + sec).toDouble() * 4.84814e-6
+    (deg * 3600 + min * 60 + sec) * 4.84814e-6
 
 /**
  * Тривиальная (1 балл)
